requirements.txt
sentence-transformers
datasets
python-dotenv
google-generativeai
